@@ -1,10 +1,17 @@
-module github.com/vakenbolt/go-test-report
+module github.com/DarkDrim/go-test-report
 
-go 1.13
+go 1.21
 
 require (
 	github.com/spf13/cobra v1.1.3
 	github.com/stretchr/testify v1.6.1
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/inconshreveable/mousetrap v1.0.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200615113413-eeeca48fe776 // indirect
 )
