@@ -159,6 +159,9 @@ func initRootCommand() (*cobra.Command, *templateData, *cmdFlags) {
 				return err
 			}
 			err = generateReport(tmplData, allTests, testFileDetailByPackage, elapsedTestTime, reportFileWriter)
+			if err != nil {
+				return err
+			}
 			elapsedTime := time.Since(startTime)
 			elapsedTimeMsg := []byte(fmt.Sprintf("[go-test-report] finished in %s\n", elapsedTime))
 			if _, err := cmd.OutOrStdout().Write(elapsedTimeMsg); err != nil {
@@ -265,7 +268,9 @@ func readTestDataFromStdIn(stdinScanner *bufio.Scanner, flags *cmdFlags, cmd *co
 func getAllDetails(listFile string) (testFileDetailsByPackage, error) {
 	testFileDetailByPackage := testFileDetailsByPackage{}
 	f, err := os.Open(listFile)
-	defer f.Close()
+	defer func() {
+		_ = f.Close()
+	}()
 	if err != nil {
 		return nil, err
 	}
@@ -307,7 +312,7 @@ func getPackageDetails(allPackageNames map[string]*types.Nil) (testFileDetailsBy
 		})
 	}
 	go func() {
-		g.Wait()
+		_ = g.Wait()
 		close(details)
 	}()
 
