@@ -1,3 +1,3 @@
 package main
 
-const version = "0.9.3"
+const version = "1.0.0"
